contribute to translation, create and import process simplification, do not show error modal for the first 10 seconds (#94)
* contribute_to_translation

* backup reminder
simplify create wallet process

* cleanup create/import
reduce number of files, remove unused

* ui_phrase.json

* do not show errors for the first N second. let node load and be responsive
