+++ src/js/api/index.js
@@ -214,7 +214,7 @@ class API {
     getStatsTransactionWithDataReceived(address_key_identifier, data_type, create_date_begin) {
         return this.fetchApiMillix(`/wWo8DCcoXVlpczoP`, {
             p6 : create_date_begin,
-            p9 : address_key_identifier.startsWith('1') ? '0a30' : 'la3l',
+            p9 : address_key_identifier?.startsWith('1') ? '0a30' : 'la3l',
             p10: address_key_identifier,
             p11: 'Adl87cz8kC190Nqc',
             p12: data_type
