+++ src/js/common/translation.js
@@ -18,30 +18,32 @@ class Translation {
     getPhrase(phrase_guid, replace_data = {}) {
         const phrase_data = this.getCurrentTranslationList().find(element => element.phrase_guid === phrase_guid && element.language_guid === this.getCurrentLanguageGuid());
         let phrase        = phrase_data?.phrase;
-        phrase            = this.htmlSpecialCharsDecode(phrase);
+        if (phrase) {
+            phrase = this.htmlSpecialCharsDecode(phrase);
 
-        let phrase_suffix = '';
-        if (sessionStorage.getItem('show_phrase_guid') === '1') {
-            phrase_suffix = ` (${phrase_guid})`;
-        }
+            let phrase_suffix = '';
+            if (sessionStorage.getItem('show_phrase_guid') === '1') {
+                phrase_suffix = ` (${phrase_guid})`;
+            }
 
-        if (!_.isEmpty(replace_data)) {
-            let result = [];
-            _.forOwn(replace_data, function(value, key) {
-                let replace_key   = `[${key}]`;
-                let result_phrase = phrase.split(replace_key);
-                if (result_phrase.length > 1) {
-                    result.push(result_phrase.shift());
-                    result.push(value);
-                    phrase = result_phrase.join('');
-                }
-            });
-            result.push(phrase);
-            result.push(phrase_suffix);
-            phrase = <>{result}</>;
-        }
-        else {
-            phrase += phrase_suffix;
+            if (!_.isEmpty(replace_data)) {
+                let result = [];
+                _.forOwn(replace_data, function(value, key) {
+                    let replace_key   = `[${key}]`;
+                    let result_phrase = phrase.split(replace_key);
+                    if (result_phrase.length > 1) {
+                        result.push(result_phrase.shift());
+                        result.push(value);
+                        phrase = result_phrase.join('');
+                    }
+                });
+                result.push(phrase);
+                result.push(phrase_suffix);
+                phrase = <>{result}</>;
+            }
+            else {
+                phrase += phrase_suffix;
+            }
         }
 
         return phrase;
