+++ src/js/components/action-view.jsx
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {withRouter} from 'react-router-dom';
 import OutputAggregateView from './action/output-aggregate-view';
 import ValidationResetView from './action/validation-reset-view';
-import MnemonicPhraseView from './action/mnemonic-phrase-view';
+import BackupView from './action/backup-view';
 
 
 class ActionView extends Component {
@@ -10,7 +10,7 @@ class ActionView extends Component {
         return (
             <>
                 <OutputAggregateView/>
-                <MnemonicPhraseView/>
+                <BackupView/>
                 <ValidationResetView/>
             </>
         );
